requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-texttospeech==2.17.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.9
joblib==1.4.2
nltk==3.9.1
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
regex==2024.9.11
requests==2.32.3
rsa==4.9
tqdm==4.66.5
urllib3==2.2.3
